src/test/java/com/leetcode/sort/hard/MaximumGap_164Test.java
package com.leetcode.sort.hard;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class MaximumGap_164Test {
  private final MaximumGap_164 solution = new MaximumGap_164();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments(new int[] {3, 6, 9, 1}, 3), arguments(new int[] {10}, 0));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void maximumGap(int[] nums, int expectedResult) {
    int actualResult = solution.maximumGap(nums);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
